+++ lib/domain/model/model_summary_activity.dart
@@ -193,7 +193,7 @@ class SummaryActivity {
     gearId = json['gear_id'];
     fromAcceptedTag = json['from_accepted_tag'];
     averageSpeed = json['average_speed'];
-    maxSpeed = json['max_speed'];
+    maxSpeed = json['max_speed'].toDouble();
     averageCadence = json['average_cadence'];
     averageWatts = json['average_watts'];
     weightedAverageWatts = json['weighted_average_watts'];
